snmplib/snmp_api: Remove the request on the session when the sending is failed
snmpd is terminated abnormally due to an invalid memory access after
the sending of a request is failed.

The time out callback for the failed request is executed when the
session is closing because the request remains in the internal session.
The cleanup for the request is executed on the
callback(NETSNMP_CALLBACK_OP_SEND_FAILED,) and also on the time out
callback(NETSNMP_CALLBACK_OP_TIMED_OUT,), so the wrong memory access
happens.

Remove the failed request from the internal session after the callback
for the failed request is done.

Signed-off-by: Masayoshi Mizuma <m.mizuma@jp.fujitsu.com>
Reported-by: Shogo Matsumoto <shogo.matsumoto@jp.fujitsu.com>